adding encd and uncd scripts
this simplifies dealing with encrypted folders at the top level of a
home directory. The folders are encrypted with encfs and stored in a
"staging dir", specified by the ~/.stagingdir file or by default
assumed to be ~/.staging/

encd looks in the staging dir for an encfs folder and gpg encrypted
file which is the encfs key. It decrypts the encfs key using gpg and
uses that to mount the encfs folder into home directory. Uncd simply
unmounts the folder and removes it from the home directory.

Using gpg means we have the protection of password protection on the
gpg key, or ability to move the gpg key onto a gpg card.